requirements.txt
aiofiles==23.2.1
fastapi==0.99.1
Jinja2==3.1.2
mangum==0.17.0
openai==0.28.1
python-dotenv==1.0.0
python-multipart==0.0.6
pinecone-client==2.2.0